pyproject.toml
[tool.poetry]
name = "uniflow"
version = "0.1.1"
description = ""
authors = ["untill99 <gabriel.kasten@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.6"
pandas = "2.2.3"
oracledb = "2.4.1"
psycopg2 = "2.9.10"
requests = "2.28.1"
apache-airflow = "2.10.1"
apache-airflow-providers-oracle = "3.11.0"
apache-airflow-providers-common-sql = "1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
build = 'poetry build'
list_dependencies = 'poetry show --tree -l'
